Rename existing `suffix_array` module to `sample`
