+++ src/search.rs
@@ -5,8 +5,8 @@ use crate::converter;
 
 use crate::converter::IndexWithConverter;
 use crate::iter::FMIndexBackend;
+use crate::sample::HasPosition;
 use crate::seal;
-use crate::suffix_array::HasPosition;
 
 /// An object containing the result of a search.
 ///
